+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # FlexMessage CHANGELOG
 
+## FlexMessage 0.0.3
+
+- Renamed `Reset()` to `Flush()`
+
+### FEATURES
+- Adding `Print()` Function
+- Implemented colors for `Print()`
+
 ## FlexMessage 0.0.2
 This new release introduces `CHANGELOG.md` :-)
 ### FEATURES